voltage
Extended the system slightly to allow 'voltage' of the flight pack to be detected.

This is a little 'odd' in implimentation.

What happens is you take a servo signal port.. and map it to a PWM input on the teensy.

You can then on opentx/jeti etc send a value by telemetry to the pwm channel.

Essentially.. take the telemetry of the flight pack from your radio telemetry.. and send it back to the receiver as a value betteen -1000 & +1000.

This is then read into the teensy on pin 10 as a pwm signal; and translated to the right voltage for display.

Its dirty.. but effective.